create separate unittest folder and add dummytest
