+++ pom.xml
@@ -55,6 +55,7 @@
 	</dependencies>
 
 	<build>
+		<testSourceDirectory>src/test/java/com/team404/kainosproject/unittests</testSourceDirectory>
 		<plugins>
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
