+++ src/test/java/com/team404/kainosproject/intergrationtests/KainosProjectApplicationTests.java
@@ -1,0 +1,13 @@
+package com.team404.kainosproject;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class KainosProjectApplicationTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
