+++ src/test/java/com/team404/kainosproject/unittests/DummyUnitTest.java
@@ -1,10 +1,10 @@
-package com.team404.kainosproject;
+package com.team404.kainosproject.unittests;
 
 import static org.junit.Assert.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class KainosProjectApplicationTests {
+class DummyUnitTest {
 
   @Test
   void testAutomaticTestsOnPushMergeToMain() {
